requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.4
openpyxl==3.1.2
schedule==1.2.0
python-dotenv==1.0.0
flask==3.1.0
flask-mail==0.9.1 